jsonrpc: add request parser with basic check
A request parser is added with basic JSON-RPC 2.0 checks on requests.
It supports simple or multiple requests (batch) and call a callback
for each request. Then, user can specify a result or an error.

By default, if no callback is specified of user callback doesn't set
any result or error, "Method not found" is returned.

Parameters and results use GVariant to pass data in order to hide
JSON objects which stay internal to melo_jsonrpc.