jsonrpc: add "required" option in parameter schema
A new option named "required" is added to the schema used to check and
convert the request parameters. If this option is present and set to
false, during check or conversion, the function won't return error
anymore.
This option is optional and can be ommited but it will, by default,
be considered as true.