jsonrpc: add an error argument to params helpers
In order to reduce code on user callback side, we add an argument
error to melo_jsonrpc_check_params(), melo_jsonrpc_get_object() and
melo_jsonrpc_get_array() functions.
When no params are provided, an Invalid request error is set and when
a bad or a missing parameter is found, an Invalid params error is set.