www: debug: get module list and info by id
This patch adds a section named Module in which we get the module list
and in which we can display the details of one listed modules (by
clicking on list item).
To call the JSON-RPC method, we use post() function from JQuery.
The JQuery 3.0.0 minimal script is also embedded in the commit.