browser: file: add volume managment
This patch adds support for volumes like flash drive or SD-card.
We can list every volumes attached to the machine thanks to GIO, and
then mount and explore the content. Every volumes are identified by
their UUID or display name. In path, the volume name is prefixed with
an underscore.