browser: file: improve volume and mount managment
To reduce code and time, the volume and mount managment has been
improved: the volume list and mount list are now merged together and
the resulting list is sorted.

Moreover, instead of using the name or UUID of the volume (or mount),
we generate a sha1 based on the GObject instance pointer. This id is
stored in data section of the GObject and is also stored in a hash
table in order to find quickly the GVolume or GMount needed for
directory listing.