file: browser: add a Network category at root
In order to explore network volumes, this patch adds a new main
category at root: Network. It uses the "network://" URI to scan all
shares available on the local network.
Since all elements in "network://" are shortcut, we convert the target
URI to sha1 and store it in a hash table in order to do not expose the
URI to user, but instead, a unique ID which is used to retrieve the
corresponding shortcut.