player: jsonrpc: add a tags entry in .get_status
A new MeloPlayerJSONRPCFields is added in order to ask for tags in
player.get_status request.
A new array of fields named "tags" is then required to ask which tags
must be retrieved.
Finally, if tags are requested, a new member called "tags" is added to
result and it carries an object containing requested tags.