tags: add a timestamp to track updates
This patch adds a timestamp got from monotonic clock in order to track
updates of MeloTags.
A new optional parameter is also added to the JSON-RPC request to set
the latest timestamp: if the timestamp is not older than the timestamp
in MeloTags, the tags member is not filled. It is very usefull when we
do a polling on player status: less data are copied and send on
network.