config: add MeloConfig, a configuration handler
A new GObject is introduced by this patch: MeloConfig. This object
will handle every settings that a user need to store. Each MeloConfig
are identified with a unique ID (like MeloModule).

A configuration is described with help of MeloConfigGroup and
MeloConfigItem structures: a configuration is composed of many groups
which are composed of many items of different types (boolean, integer,
double, string, ...). It follows the same schema that GKeyFile, with
groups and key / value pairs;

Values are stored separatly from the descriptor, in an array of
MeloConfigValues containing an array of MeloConfigValues and a hash
table to find quickly a value from its ID.