config: jsonrpc: add a config.get method
A new JSON-RPC method is added with this patch, in order to get full
configuration from a MeloConfig. The method takes as main parameter,
the ID of MeloConfig requested and an optional group ID to get only
a part of the complete configuration.
The result is an array containing all groups and their items and
values.