config: add configuration set/update support
This patch adds a new helper to update values of a MeloConfig object,
with the introduction of two tables handling new values and update
state of each values. It uses also two new callbacks to notify the
developer for a check of new values and / or availability of updated
values.

In addition, a new JSON-RPC method named config.set is added to send
new values to Melo.

It also fix some bugs on lookup of IDs in hash table: the pointer was
not correctly handled and some troubles with stack was occuring.