airplay: improve support of RTSP server
Now, the RTSP server repond to each required steps of the Airplay
protocol in order to setup a new connection.
This patch adds use of Gstreamer SDP parser to read ANNOUNCE requests
of Airplay and to get AES key which will be used for the following
TCP/RTP session.
It also adds a MeloPlayerAirplay which only create an empty pipeline
which only instanciate a TCP server source or a UDP source linked to a
fake sink. It allows to finalize all RTSP transactions.