airplay: add PCM support for RAOP streams
This patch adds support for PCM stream on AirTunes, by adding an
encoding-name capability to rtpraopdepay element which permits to pass
the codec name (L16, ALAC or AAC) to the element.

HACK: after some tests on an iPod Touch 2nd gen with iOS 4, when PCM is
used, a strandard ALAC stream is received, so a define is available to
enable a piece of code to decode PCM as ALAC stream (enabled by
default).