upnp: add a new MeloModule for UPnP / DLNA support
This new MeloModule named MeloUpnp is a module which brings support of
UPnP AV and DLNA protocols, with a player which acts as a
MediaRenderer. All tasks are managed by Rygel build on GUPnP and
Gstreamer.
Later, a browser will be added in order to navigate through UPnP
servers available on network.