libmelo: browser: add tags support in get_list
Two new arguments are added with this patch to the MeloBrowser
get_list callback: tags_mode and tags_fields. The first one is to
specify which mode must be used with Tags retrieving (no tags, cached
tags, all tags with caching of not cached tags, ...) and the second
one to specify which tags must be retrieved.

In addition, the JSON-RPC method named browser.get_list is updated and
use a new node name tags in which mode and fields can be passed.