file: browser: add tags support to get_list
This patch adds support of MeloTags in get_list callback of
MeloBrowserFile. It basically (when asked through the tags mode) tries
to load tags from database and generate the requested MeloTags for
each files. If no records are found in database, in case of CACHED
mode, the URI file is send to the main Gstdiscoverer of object and
tags are retrieved and added to database asynchronously. In case of
one of FULL modes, the tags are retrieved synchronously and call can
then take more time to return.