libmelo: browser: add go feature in info
This patch introduces a new feature named go: an input form to go
directly to a path or URI within MeloBrowser. Basically, it is only a
part of MeloBrowserInfo which tells if MeloBrowser instance support go
feature and which text to display to the final user.
It also adds a new fields and an object definition in JSON-RPC
interface.