file: browser: add get_cover for cover URL support
This patch adds URL cover support for MeloBrowserFile. All cover arts
saved on disk are now read into get_cover and send to HTTP client. The
URL is generated in MeloFileDB when MeloTags are created from a find
operation and the URL is generated in MeloBrowserFile when gstreamer
tags discoverer is used and a new media is added to database.