radio: player: add support for cover art / logo
The MeloTags provided by MeloBrowser during melo_player_play call is
saved internally and used as base for MeloPlayerStatus. When a
MeloTags is generated from audio stream, the cover of current playing
song is set to MeloPlayerStatus or, if not available, the cover
(station logo / current song) provided by MeloBrowser is used.