libmelo: browser: add token support for navigation
Some MeloBrowser depend on API which use token or handle to navigate
through results of listing or research. To handle this behavior, we
add a prev_token and a next_token to the MeloBrowserList which is
passed to sub layer as JSON-RPC interface.

This patch also adds a new node named 'token' in which user can pass
token for navigation instead of offset.