libmelo: playlist: simple: add get_cover support
A basic get_cover implementation is added in MeloPlaylistSimple which
get GBytes from a playlist item MeloTags.
A new function is also added to enable or not the cover URL override
during add process: if a cover is present and cover URL override is
enabled, MeloPlaylistSimple set cover URL with its instance. It can
be useful when MeloTags comes from a MeloPlayer or a MeloBrowser in
which the cover URL has been set with a binding which always downloads
the cover art from a HTTP server.