player: add volume support and control
This patch adds two new callbacks to MeloPlayer in order to get and
control volume of a player. It also adds a new field named volume in
MeloPlayerStatus.
Finally, it adds a new JSON-RPC method named player.set_volume to
control volume of the player. A new json node is also added in
response of player.get_status method.