melo: discover: add retry of device registration
When inernet connection is not available at start or when user enables
Melo device registration in settings, the device was not added to Melo
website and all next events on interfaces were discarded even if one
of these events was related to one interface associated to internet
connection. Now, the complete device registration is retried at any
interfaces event until the device has been registered successfuly.
Additionally, the max concurent connection on SoupSession has been
drop to one in order to guarantee that device is registered before
adding interfaces.