libmelo: player: add loading and buffering states
This patch adds two new MeloPlayerState which are LOADING and
BUFFERING to respectively indicates when player is loading a media
(just after call to play and before playing first sample of audio) and
when player is buffering next data (playback is paused during
buffering so user can be aware of current operation). The buffering
is accompanied by a new field in MeloPlayerStatus to give current
amount of data in buffer (in percent).

It also updates the JSON-RPC methods according to new states and
fields brang by this patch.