libmelo: player: add a name property to MeloPlayer
A new property 'name' is added to MeloPlayer in order to give a name
to a MeloPlayer instance, in addition to its ID. It is very useful for
GUI which can need to display a name or a title for each players.
It also renames melo_player_get_name() function used to get name field
from MeloPlayerStatus, into melo_player_get_media_name(). Now, the
melo_player_get_name() is used to get MeloPlayer name instance.