melo: httpd: add /version handler
A new path handler for '/version' is added to check if a Melo instance
is available on network.
When instance is allowed to be registered on Melo website, its IP
address and port are saved on the website, which allows the user to
find easily its Melo instance on local network. To help more, the new
'/version' handler allows Melo website to check if Melo instance is
started and available on network, and if not, to notify the user about
its unavailability.