file: library; implement search feature
This patch adds the search callback implementation to MeloLibraryFile
class to allow search in media library.

To perform an easy and performant search, the MeloFileDB class is
reworked to uses the Full-text searches module of SQLite. It adds four
new virtual table 'song_fts', 'artist_fts', 'album_fts' and
'genre_fts' to stores the strings of respective 'song', 'artist',
'album' and 'genre' tables.
It allows to search in media database with only one string for each
MeloFileDBFields and all power of the Full-text searches feature.

The MeloLibraryFile search feature works only for a media file: the
artist, album, genre, date and track(s) number is not yet supported.