lib: browser: update function parameters
This patch changes the parameters for melo_browser_get_list(),
melo_browser_search(), melo_browser_add() and melo_browser_play(), and
their respective callback to reduce the parameter count and avoid
changing all implementation prototypes in derivated classes.

Now, the functions take a string and a parameters structure named
respectively MeloBrowserGetListParams, MeloBrowserSearchParams,
MeloBrowserAddParams and MeloBrowserPlayParams.

For the add and play feature, a token string is added to parameters
to identify a list when path represents a list and not a unique item.
The JSON-RPC methods are updated according to this new parameters
specification.