melo: add HTTPS server support
This patch adds the SSL support to HTTP server for serving secured
pages. A secured port is added to main configuration to handle the
port for HTTPS instance.
This port is set to 0 by default, which disable the feature.
If the port is set to non-zero value, the instance is enabled and a
self-signed certificate is generated if it doesn't exist.