lib: plugin: add an API version check
An API version field is added to MeloPlugin descriptor to embed the
API version on which a plugin has been built.
Now, if the API version of a plugin and the Melo library mismatch, the
plugin is not loaded.
An helper is added to definition to update easily code of all plugins
source code.