file: library: mount volume when needed
This patch moves the code for mounting remote file system into a new
file name melo_file_utils.c which is shared between the file browser
and the library browser.
Now; when a file is added or played from library browser and it is on
a not mounted volume, it is done automatically.