lib: browser: fix item types and actions
Instead of using a simple string to identify the type of an item in
the MeloBrowserItemList, we use now an enumerator which fix propose a
list of standard item types (media, file, folder, device, ...).
In addition, the 'add' and 'remove' fields are replaced by a bit-field
which list the possible actions of an item (play, add, remove, ...).
Now, the actions on an item will be performed through a unique
function named melo_browser_action(). It allows to add more action in
future development and simplify browser implementation.
This patch also updates the JSON-RPC interface (by reducing to only
one browser.action method) and MeloBrowser implementation for file and
radio modules.