lib: playlist: rename item fields
The 'name' and 'full_name' fields of the MeloPlaylistItem structure
are respectively renamed to 'id' and 'name' in order to be more
coherent with the MeloBrowserItem structure. The 'name' parameter of
all MeloPlaylist functions are also renamed to 'id', and the
subclasses are updated according to the new MeloPlaylist API.
It also updates the JSON-RPC methods by replacing 'name' by 'id' in
JSON-RPC response and 'media_id' in JSON-RPC requests, and 'full_name'
by 'name'.