lib: tags: rework of cover handling
In order to optimize memory usage and access to the cover data
attached to a Melotags, the internal system of MeloTags has been
reworked. In addition, the URL system has been also modified for
easier development of MeloBrowser, MeloPlayer and MeloPlaylist.

Now, a cover can be added to a MeloTags with a GBytes containing the
image data or a simple URL. A unique ID is then generated and stored
in the MeloTags 'cover' field. This ID point to the image data or the
URL provided during cover set.
An internal cache stores all cover image and shares all duplicated
images or URL thanks to the unique ID generated on the data or URL
provided. It prevents doing many copy of the same image in memory.
If an URL is set, the image data is downloaded only at first call to
functions intended to retrieve the image data, and are then stored
internally. It means that image download is not anymore needed in
object implementations and the many _get_cover() callbacks are not
used anymore.

A cover can be set as 'non-persistent' which means it will be freed
during free of the last dependant MeloTags, set as 'semi-persistent'
which means it will be kept internally until end of the program or set
as 'persistent' which means it will be stored on disk and will be
available at next startup of the application.

Thanks to the unique ID, two functions are provided to get the cover
data from this ID, which removes all dependencies to the GObject
which set the cover to the MeloTags. That is why the _get_cover()
callbacks are removed by this patch as the cover URL field of the
MeloTags.

Many objects are updated in this patch due to the dependencies to the
MeloTags API.

Finally, it add the documentation to the MeloTags and the cover cache
system.