libmelo: sink: save volume/mute into file
A new mechanism is added to MeloSink in order to save current settings
into a file to restore them at next boot.
The values are saved to the file with an interval of 10s in order to
avoid a huge write operation when the volume is modified through a
slider on user interface.