file: player: decode only audio tracks
This patch forces the uridecodebin used in file player implementation
to decode and play only audio tracks when the media is a video (by
default uridecodebin decodes all streams even if source pad is not
used).
The trick is to force the "caps" property of the element to
"audio/x-raw(ANY)".