melo: add network support
The network devices configuration such as Ethernet or WiFi devices is
added to Melo, in order to let the user to change IP settings, select
a WiFi access point to connect to or setup a WiFi hotspot.
A dedicated websocket and its protocol buffer file are defined in Melo
source program.