lib: library: add flags and favorite feature
A new parameter is added to the melo_library_add_media() to set
specific flags to a media and then extend easily the library database
system. An helper is added to set / unset one or more flags from a
media.

The first flags added are FAVORITE and FAVORITE_ONLY, to bring a new
interesting feature which allows users to save their favorite medias
in the library database. The FAVORITE_ONLY is intended for browser
which want to save only favorites in database, and then the media is
purely deleted from the library database if the user remove the mark.
This flag will also be added to player and playlist class to set the
current playing media as favorite.