modules: file: browser: add favorite support
The favorite status is now retrieved from the library and set in the
media item response. Additionally, the favorite related actions are
added to media actions to set / unset the 'favorite' marker from a
media.

This patch fixes also some issue on filter which was missing for
actions and the path used with melo_library_find() in actions.