modules: add UPnP module for UPnP / DLNA rendering
A new optional module is added to Melo environment in order to support
UPnP / DLNA features.
Currently, it implements only a renderer based on Rygel, which let
Melo to appear on network as UPnP / DLNA renderer, as a smart TV could
do.
Later, a browser will be added to let user to browse and play medias
from UPnP / DLNA servers on the network.