lib: playlist: rework depth support
The playlist is completely reworked internally to a depth greater to
one nested media list: now, it handles dynamically a deeper depth and
allows to create complex playlist trees.

The move and delete is also added to the playlist in order to let user
to move some medias from on sub-list to another one and to delete
easily some unwanted medias.

Finally, the automatic playlist control is improved to handle some
specials cases, and move forward when a play request is done on an
unplayable media.