modules: file: browser: implement scan action
The 'scan' custom action is now available and it will scan recursively
the folder pointed by the path field of the request, and it will fill
the media library with detected files, with their respective tags.

In addition, the tags discoverer usage is improved to handle correctly
the cancellation of a request (before, the request was destroyed only
after completion of the discoverer).