modules: file: browser: handle removable devices
This patch adds the volume and mount handling from the GIO to the file
browser, in order to let user mount, browse and unmount / eject
removable devices such as USB flash drives or remote NAS.

A volume monitor is instantiated to add / remove volume and mounts
automatically in the browser. An 'added' and 'removed' events could be
generated from the callbacks.