lib: rework of request API and related
The request object is now an independent object and then it do not
rely on the source object. The melo_browser_handle_request() will
return a MeloRequest reference which can be canceled later easily
with melo_request_cancel().

For browser implementations, the request completion must be done
explicitly with melo_request_complete(). The final calls to
melo_request_unref() can be replaced with this new function.

The playlist cancellation function is removed since it was unused.