proto: browser: rework actions and events
The available actions for the current Browser.Request.get_media_list
are now listed on Browser.Response.MediaList and the applicable
actions are now described with a list of IDs referring to the
available actions list. It reduces the size of the message and then
reduce processing time during serialization / de-serialization.

Additionally, the Browser.Request.DoAction has a params field to hold
some complementary parameters to the action request.

Finally, the action and event list is filled with some actions related
to file browser and manipulation (create, rename, move, paste,
delete).